==> 1_globe_image_scraping_processing.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup
import os

df = pd.read_csv('data/country_list.csv')

df.head()

headers = {
    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)'
}

countries = df['country']

for country in countries:
    # country_encoded = quote(country.replace(" ", "_"))
    file_name = f"{country}_(orthographic_projection).svg"
    url = f"https://en.wikipedia.org/wiki/File:{file_name}"

    try:
        html = requests.get(url, headers=headers).text
        soup = BeautifulSoup(html, 'html.parser')
        con = soup.find('div', id='file')
        if not con:
            print(f"[{country}] - Image container not found.")
            continue
        
        globe_tag = con.find('img')
        if not globe_tag or not globe_tag.get('src'):
            print(f"[{country}] - Image tag or src not found.")
            continue
        
        globe_url = 'https:' + globe_tag['src']
        globe_name = country.replace(" ", "_") + '.png'
        
        os.makedirs('globe_image', exist_ok=True)
        img_response = requests.get(globe_url, headers=headers)
        
        if img_response.status_code == 200:
            with open(os.path.join('globe_image', globe_name), 'wb') as file:
                file.write(img_response.content)
            print(f"[{country}] - File downloaded.")
        else:
            print(f"[{country}] - Failed to download image.")
    
    except Exception as e:
        print(f"[{country}] - Error: {e}")


folder_path ='globe_image'
file_names = os.listdir(folder_path)
image_files = [f for f in file_names if f.endswith('.png')]

df_globe = pd.DataFrame(image_files, columns=['image_globe'])

df_globe['image_globe'].str.split('.')[0][0]

df_globe.head(5)

df_globe['country'] = df_globe['image_globe'].apply(lambda x: x.split('.')[0])

df_globe.head()

df_globe.loc[:,'country']= df_globe['country'].str.replace('_',' ')

df_globe.tail(30)

df_globe.to_csv('data/df_globe.csv')
