==> 3_country_info_scraping_processing.ipynb <==
import pandas as pd

df = pd.read_csv('data/country_list.csv')

df.head()

#✅✅✅✅✅✅✅✅✅✅✅✅✅
import requests
from bs4 import BeautifulSoup
import re

countries = df['country']
country = []
info = []
for i in countries:
    url = f'https://en.wikipedia.org/wiki/{i.replace(" ", "_")}'

    country.append(i)
    # Step 1: Request the page
    response = requests.get(url, timeout=10)
    response.raise_for_status()
    
    # Step 2: Parse HTML
    soup = BeautifulSoup(response.text, 'html.parser')
    
    # Step 3: Get the main content div
    content_div = soup.find_all('div', class_='mw-content-ltr mw-parser-output')
    
    list = []
    for i in content_div:
        list.append(i.find_all('p'))
    main_para = list[0][1].text
    clean_text = re.sub(r'\[\w* ?\d*\]', '', main_para)
    info.append(clean_text)
    
# df_info = pd.DataFrame(info, columns=["details"])

df_info = pd.DataFrame({
    'country': country,
    'details': info
})

df_info.tail()

# Save in csv file
df_info.to_csv('data/df_info_temp.csv')

pd.set_option('display.max_colwidth', None)

df_info.head(2)

df_info.loc[:,'details'] = df_info['details'].str.replace('\n','').str.strip()

df_info.tail(50)

df_info.to_csv('data/df_info.csv')
