==> 4_merge_all_scraped_data_processing.ipynb <==
import pandas as pd

country_list = pd.read_csv('data/country_list.csv')
df_globe = pd.read_csv('data/df_globe.csv')
df_flag = pd.read_csv('data/df_flag.csv')
df_info = pd.read_csv('data/df_info.csv')
df_level = pd.read_csv('data/df_level.csv')

df_level.head()

country_list = country_list.drop(columns='Unnamed: 0')

df_globe = df_globe.drop(columns='Unnamed: 0')

df_flag = df_flag.drop(columns='Unnamed: 0')

df_info = df_info.drop(columns='Unnamed: 0')

df_level = df_level.drop(columns='Unnamed: 0')

df = country_list.merge(df_globe, how='left', on='country')

df = df.merge(df_flag, how='left', on='country')

df = df.merge(df_info, how='left', on='country')

df = df.merge(df_level, how ='left', on ='country')

df.head()

df = df.rename(columns = {'image_name':'image_flag'})

pd.set_option('display.max_colwidth', None)

df.tail(2)

df.info()

df.head()

df[df.duplicated()]

df.drop_duplicates(inplace = True)

df[df['country'].duplicated()]

df.info()

df[df['image_flag'].isnull()]

df.to_csv('data/df_country_info_all.csv')
